Delete function added to the application
